requirements.txt
fastapi
uvicorn
torch
pillow
numpy
torchvision

